flask-backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
firebase-admin==6.0.1
python-dotenv==1.0.0
